Implemented player movement, and other changes. See description for details.
Player movement now uses stamina and updates an on-screen slider accordingly. Once the player is out of stamina, they can not move.

PlayerHealthManager is now obsolete as the GameManager subscribes to the player's OnDeath event on its own.

The GameManager is no longer a singleton (for now :wink: ). I found it unnecessary since I'm using events.