More work on duelist and shotgunner enemies. Changed how shooting over cover works.
Duelist is pretty much implemented, and the shotgunner enemy works, but is general all AI that has to move around the map occasionally get stuck on obstacles they shouldn't. It's something to do with their capsule colliders. I'll experiment with this in mind.

As for shooting over cover, I changed it so whenever an enemy or a player is in range of a cover object, their bullet won't collide with that cover object.